consolidate user/journals
