+++ mobile/Features/Feed/Screens/SignUp.jsx
@@ -134,12 +134,12 @@ export default function SignUp({ navigation }) {
         email: userEmail,
         password,
       };
-      const res = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/users/signup`, userData);
+      const res = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/signup`, userData);
 
       if (res.data.error) {
         console.error(res.data.error);
       } else { // If sign up is successful
-        const res2 = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/users/signin`, userData);
+        const res2 = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/signin`, userData);
         // Sets current state variables for session
         dispatch(login(res2.data));
         setEmail('');
