+++ mobile/Features/Onboarding/Screens/Login.jsx
@@ -31,7 +31,7 @@ export default function Login({ navigation }) {
         email: userEmail,
         password,
       };
-      const res = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/users/signin`, userData);
+      const res = await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/signin`, userData);
       if (res.data.error) {
         console.error(res.data.error);
       } else {
