+++ server/models/OfficialUserSchema.js
@@ -1,30 +1,5 @@
 const mongoose = require('mongoose');
-
-const checkInPreferenceSchema = new mongoose.Schema({
-  category1: {
-    type: [String],
-    default: [],
-  },
-  category2: {
-    type: String,
-  },
-  category3: {
-    type: String,
-  },
-});
-
-const notificationPreferenceSchema = new mongoose.Schema({
-  timeOfDay: {
-    type: String,
-    enum: ['morning', 'afternoon', 'evening', 'night'],
-    default: 'morning',
-  },
-  recurrence: {
-    type: String,
-    enum: ['daily', 'weekly', 'monthly'],
-    default: 'daily',
-  },
-});
+const bcrypt = require('bcrypt');
 
 const officialUserSchema = new mongoose.Schema({
   firstName: {
@@ -32,11 +7,11 @@ const officialUserSchema = new mongoose.Schema({
     type: String,
   },
   lastName: {
-    required: true,
+    required: false,
     type: String,
   },
   username: {
-    required: true,
+    required: false,
     type: String,
   },
   password: {
@@ -46,6 +21,7 @@ const officialUserSchema = new mongoose.Schema({
   email: {
     required: true,
     type: String,
+    unique: true,
   },
   profilePicture: {
     required: false,
@@ -64,17 +40,37 @@ const officialUserSchema = new mongoose.Schema({
     type: [Object],
   },
   checkInPreferences: {
-    default: undefined,
-    type: checkInPreferenceSchema,
+    default: {},
+    type: Object,
   },
-  notificationPreferences: {
-    default: undefined,
-    type: notificationPreferenceSchema,
+  pushNotifs: {
+    default: {
+      time: null,
+      reminders: [],
+    },
+    type: Object,
   },
   ChallengeDay: {
     default: 0,
     type: Number,
   },
 });
+
+// Hashes password upon creation of user
+officialUserSchema.pre('save', async function hashPass(next) {
+  const user = this;
+  if (!user.isModified('password')) return next();
+  const hash = await bcrypt.hash(user.password, 10);
+  user.password = hash;
+  next();
+  return null;
+});
+
+// Check if password is correct
+officialUserSchema.methods.isValidPassword = async function check(password) {
+  const user = this;
+  const compare = await bcrypt.compare(password, user.password);
+  return compare;
+};
 
 module.exports = mongoose.model('OfficialUser', officialUserSchema);
