--- server/models/UserSchema.js
@@ -1,39 +1,0 @@
-const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
-
-// Example of a model schema to validate and structure documents
-const userSchema = new mongoose.Schema({
-  email: {
-    required: true,
-    type: String,
-    unique: true,
-  },
-  password: {
-    required: true,
-    type: String,
-  },
-  firstName: {
-    required: true,
-    type: String,
-  },
-});
-
-// Hashes password upon creation of user
-userSchema.pre('save', async function (next) {
-  const user = this;
-  if (!user.isModified('password')) return next();
-  const hash = await bcrypt.hash(user.password, 10);
-  user.password = hash;
-  next();
-});
-
-// Check if password is correct
-userSchema.methods.isValidPassword = async function (password) {
-  const user = this;
-  const compare = await bcrypt.compare(password, user.password);
-  return compare;
-};
-
-const User = mongoose.model('User', userSchema);
-
-module.exports = User;
