+++ server/package.json
@@ -16,11 +16,12 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "jsonwebtoken": "^9.0.2",
+    "mongodb": "^6.3.0",
+    "mongoose": "^7.6.3",
     "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
-    "mongodb": "^6.3.0",
-    "mongoose": "^7.6.3",
     "react-native-timer-picker": "^1.5.3"
   },
   "devDependencies": {
