--- server/routes/journalRoute.js
@@ -1,10 +1,0 @@
-const express = require('express');
-
-const journalRouter = express.Router();
-const journalController = require('../controllers/journalController');
-
-journalRouter.post('/createJournal', journalController.createJournal);
-
-journalRouter.get('/getAllJournals', journalController.getAllJournals);
-
-module.exports = journalRouter;
