+++ server/routes/offUserRoute.js
@@ -3,6 +3,18 @@ const express = require('express');
 const offUserRouter = express.Router();
 const offUserController = require('../controllers/offUserController');
 
+offUserRouter.post('/signin', offUserController.signInUser);
+offUserRouter.post('/signup', offUserController.signUpUser);
+
+// Protects the routes below with middleware (requires authorization header aka signed in user)
+offUserRouter.use(offUserController.authenticatePassport);
+offUserRouter.post('/getData', offUserController.getUserData);
+
+offUserRouter.get('/get', (req, res) => {
+  res.send('Get API');
+});
+offUserRouter.get('/protected', offUserController.authenticatePassport, offUserController.welcomeUser);
+
 offUserRouter.post('/createUser', offUserController.createUser);
 
 offUserRouter.get('/getUserById', offUserController.getUserById);
@@ -20,5 +32,7 @@ offUserRouter.post('/getUserChallengeDay', offUserController.getUserChallengeDay
 offUserRouter.post('/resetChallengeDay', offUserController.resetChallengeDay);
 
 offUserRouter.post('/increaseChallengeDay', offUserController.increaseChallengeDay);
+
+offUserRouter.post('/readSpecifiedFields', offUserController.readSpecifiedFields);
 
 module.exports = offUserRouter;
